created components + sites with leagues
