+++ app.vue
@@ -1,7 +1,12 @@
+<script setup lang="ts">
+import type AppBar from './components/AppBar.vue';
+</script>
+
 <template>
   <NuxtLayout>
     <v-app>
       <v-main>
+        <AppBar />
         <Navigation />
         <NuxtPage />
       </v-main>
