+++ pages/leagues/pol-1.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <span>
+      Ekstraklasa
+    </span>
+  </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style>
+
+</style>
